Adapted the sizes of memory info structure to unsigned long, to accomodate - once more - the absolutely ridiculously monstrous amount of memory the fucking Altix UV has. FFS
git-svn-id: svn+ssh://voms.jku.austriangrid.at/var/subversion/proc@28 593b3eb8-86d1-0410-841a-be20d09665e6

+++ def.h
@@ -33,12 +33,12 @@ typedef struct pidstat
 
 typedef struct meminfo
 {
-	int	memtotal;
-	int	memfree;
-	int buffers;
-	int cached;
-	int swaptotal;
-	int swapfree;
+	unsigned long	memtotal;
+	unsigned long	memfree;
+	unsigned long	buffers;
+	unsigned long	cached;
+	unsigned long	swaptotal;
+	unsigned long	swapfree;
 }mstat, *pmstat;
 
 int compare_elements ( const void *first, const void *second );
